Created Export-ADUsers.ps1
Gets the names, email addresses, and statuses of members of a supplied AD group or OU.
Interactive menus take the choice of acition and name of the target group or OU.
Any arguments simplied during invocation are ignored.
If an input contains whitespaces (e.g. Research and Development) use quotation marks "" (i.e. "Research and Development").
Outputs a CSV file contaiing the names, email addresses, and statuses of all users in the target group or OU.