Created Get-ADUserEmailAddress.ps1
Gets the names and email addresses of members of a supplied AD group.
Takes the name of the target groups as a one-word argument.
Any other arguments are simply ignored.
If the target group is has a multi-word name (e.g. Research and Development) use "" (i.e. "Research and Development").
Outputs a CSV file "C:\user_emails.csv" contaiing the name and email address of all users in the target group.